# DOCS/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Binar Academy Post API - FSW1",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://idea-academy.up.railway.app/"
    }
  ],
...
  "paths": {
...
    "/api/v1/current-user": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get current user information",
        "description": "Get current user information",
        "operationId": "currentUser",
        "security": [
          {
            "api_key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/OK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Login successfully."
                        },
                        "data": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/FAIL"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "example": "Please input email and password."
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Invalid token or JWT expired",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/FAIL"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "enum": [
                            "invalid token",
                            "jwt expired"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/register": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Register new admin",
        "description": "Register new admin",
        "operationId": "registerAdmin",
        "security": [
          {
            "api_key": []
          }
        ],
        "requestBody": {
          "description": "Register new admin",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdmin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Data was created successfully."
                    },
                    "data": {
                      "$ref": "#/components/schemas/User2"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Body is missing."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/FAIL"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "example": "Unauthorized"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Invalid operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/FAIL"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "example": "Failed to create data. Validation error"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "OK": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "OK"
          },
          "message": {
            "type": "string",
            "example": "Success"
          }
        }
      },
...
      "FAIL": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "FAIL"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "RegisterAdmin": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "TESTER"
          },
          "email": {
            "type": "string",
            "example": "test@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "test"
          },
          "phoneNumber": {
            "type": "string",
            "example": "082112345678"
          },
          "address": {
            "type": "string",
            "example": "Earth"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "d2d6d792-fe75-4613-b1c2-579ed0564b3c"
          },
          "name": {
            "type": "string",
            "example": "root"
          },
          "email": {
            "type": "string",
            "example": "root@gmail.com"
          },
          "encryptedPassword": {
            "type": "string",
            "example": "$2b$10$CZWqOKzxtogiTgPNLmxMXORb0fEuZxr0qrEuh2exgpe5I8c6Zq.Du"
          },
          "phoneNumber": {
            "type": "string",
            "example": "082112345678"
          },
          "address": {
            "type": "string",
            "example": "Earth"
          },
          "role": {
            "type": "string",
            "enum": [
              "ROOT",
              "ADMIN",
              "MEMBER"
            ],
            "example": "ROOT"
          },
          "country": {
            "type": "string",
            "example": "indonesia"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-10-18T13:42:59.162Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-10-18T13:42:59.162Z"
          }
        }
      },
      "User2": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "640f98ab-951b-4b7b-845d-6e78b411b566"
          },
          "name": {
            "type": "string",
            "example": "TESTER"
          },
          "email": {
            "type": "string",
            "example": "test@gmail.com"
          },
          "encryptedPassword": {
            "type": "string",
            "example": "$2b$10$P0VY83faE3Wy71l5ZGq4ZunVoxb6wy4q1SxZAZLU7PgOm4RMfw0oK"
          },
          "phoneNumber": {
            "type": "string",
            "example": "082112345678"
          },
          "address": {
            "type": "string",
            "example": "Earth"
          },
          "role": {
            "type": "string",
            "enum": [
              "ROOT",
              "ADMIN",
              "MEMBER"
            ],
            "example": "ADMIN"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-10-18T13:42:59.162Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-10-18T13:42:59.162Z"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}